Move setup root to screen dir.
